removed old model + mapping. commented out broken code for the moment
