+++ src/main/java/org/molgenis/bbmri/eric/controller/EricDirectoryManagerController.java
@@ -3,13 +3,11 @@ package org.molgenis.bbmri.eric.controller;
 import static org.molgenis.bbmri.eric.controller.EricDirectoryManagerController.URI;
 
 import org.molgenis.bbmri.eric.service.EricDownloadService;
-import org.molgenis.bbmri.eric.service.EricDownloadService.DownloadReport;
 import org.molgenis.ui.MolgenisPluginController;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
 
 @Controller
 @RequestMapping(URI)
@@ -32,10 +30,10 @@ public class EricDirectoryManagerController extends MolgenisPluginController
 		return "view-ericdirectorymgr";
 	}
 
-	@RequestMapping(value = "/download", method = RequestMethod.POST)
-	@ResponseBody
-	public DownloadReport download()
-	{
-		return ericDownloadService.downloadSourcesOnDemand();
-	}
+	// @RequestMapping(value = "/download", method = RequestMethod.POST)
+	// @ResponseBody
+	// public DownloadReport download()
+	// {
+	// return ericDownloadService.downloadSourcesOnDemand();
+	// }
 }
